To rpn (#7)
* Merge to_rpn to master

+++ src/calculator.c
@@ -1,4 +1,6 @@
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 #include "eval.h"
 #include "utils.h"
@@ -21,7 +23,7 @@ int main(void)
     puts(variable_request);
 
     /* TODO: Using get_variable instead */
-    for (size_t counter = 0; fgets(line, MAX_LENGTH, stdin); counter++) {
+    for (int counter = 0; fgets(line, MAX_LENGTH, stdin); counter++) {
         puts(variable_request);
         whitespace_cleaner(line);
 
@@ -37,7 +39,7 @@ int main(void)
 
     /* Use TokenArray instead char** */
     // char **tokens = calloc(MAX_LENGTH, sizeof(char*));
-    // for(size_t i = 0 ; i < 11 ; i++)
+    // for(int i = 0 ; i < 11 ; i++)
     // {
     //    *(tokens + i) = calloc(1, sizeof(char) * MAX_LENGTH);
     // }
